move explore route up out of team route
