+++ apps/editor.planx.uk/src/routes/_authenticated/app/explore.tsx
@@ -3,10 +3,10 @@ import { hasFeatureFlag } from "lib/featureFlags";
 import Explore from "pages/Explore";
 import React from "react";
 
-export const Route = createFileRoute("/_authenticated/app/$team/explore")({
-  beforeLoad: ({ params }) => {
+export const Route = createFileRoute("/_authenticated/app/explore")({
+  beforeLoad: () => {
     if (!hasFeatureFlag("EXPLORE")) {
-      throw redirect({ to: "/app/$team", params: { team: params.team } });
+      throw redirect({ to: "/app" });
     }
   },
   component: Explore,
